% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acy.R
\name{acgnlsobj}
\alias{acgnlsobj}
\title{AC GNLS Objective Function}
\usage{
acgnlsobj(x, y, tp, ga, p, la, q)
}
\arguments{
\item{x}{Concentration.}

\item{y}{Desired activity level.}

\item{tp}{Gain-loss Model Top Parameter.}

\item{ga}{Gain AC50.}

\item{p}{Gain power.}

\item{la}{Loss AC50.}

\item{q}{Loss power.}
}
\value{
Difference between GNLS model repsone at x and y.
}
\description{
GNLS objective function set to y for gnls solver.
}
