% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcempirical.R
\name{calcempirical}
\alias{calcempirical}
\title{Empirically locate top and top concentration}
\usage{
calcempirical(conc, ps, fname)
}
\arguments{
\item{conc}{Vector of concentrations}

\item{ps}{Vector of parameters, must be in order: a, tp, b, ga, p, la, q, er}

\item{fname}{Model function name (equal to model name except hill which
uses "hillfn")}
}
\value{
A named list of the top and top concentration. The elements of the list are:
 \itemize{
   \item top - empirical top
   \item x_top - concentration at which empirical top occurs
 }
}
\description{
Locate the top and the concentration at which top occurs empirically by sampling points from the experimental concentration values.
}
\examples{
conc = c(.03,.1,.3,1,3,10,30,100)
ps = c(1,2,1,2,2)
fname = "gnls"
calcempirical(conc, ps, fname)
}
