% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplObj.R
\name{cnst}
\alias{cnst}
\title{Constant Model - returns zero}
\usage{
cnst(ps, x)
}
\arguments{
\item{ps}{Vector of parameters (ignored)}

\item{x}{Vector of concentrations (regular units)}
}
\value{
Vector of model responses
}
\description{
\eqn{f(x) = 0}
}
\examples{
cnst(1,1)

}
