% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concRespCore.R
\name{concRespCore}
\alias{concRespCore}
\title{Concentration Response Core}
\usage{
concRespCore(
  row,
  fitmodels = c("cnst", "hill", "gnls", "poly1", "poly2", "pow", "exp2", "exp3", "exp4",
    "exp5"),
  conthits = T,
  aicc = F,
  force.fit = FALSE,
  bidirectional = TRUE,
  verbose = FALSE,
  do.plot = F
)
}
\arguments{
\item{row}{A named list that must include:
\itemize{
  \item conc - list of concentrations (not in log units)
  \item resp - list of corresponding responses
  \item bmed - median of noise estimate.
  \item cutoff - noise cutoff
  \item onesd - 1 standard deviation of the noise (for bmd calculation)
}
Other elements (usually identifiers, like casrn) of row will be attached to
the final output.}

\item{fitmodels}{Vector of model names to use.}

\item{conthits}{conthits = T uses continuous hitcalls, otherwise they're
discrete.}

\item{aicc}{aicc = T uses corrected AIC to choose winning method; otherwise
regular AIC.}

\item{force.fit}{If TRUE force the fitting to proceed even if there are no points
outside of the bounds (default FALSE)}

\item{bidirectional}{If TRUE allow fitting to happen in both directions (default TRUE)}

\item{verbose}{If TRUE, write extra output from tcplFit2 (default FALSE)}

\item{do.plot}{If TRUE, create a plot in the tcplFit2 function (default FALSE)}
}
\value{
One row dataframe containing all CR output and statistics and any
  identifiers from row.
}
\description{
Core of concentration response curve fitting for pvalue based cutoff. This
function calls httrFit to get curve fits, chooses the winning model, extracts
the top and ac50, computes the hitcall, and calculates bmd/bmdl/bmdu among other
statistics. Nested model selection is used to choose between poly1/poly2, then
the model with the lowest AIC (or AICc) is declared the winner. Continuous
hitcalls requires tcplFit2 to be run with force.fit = T and "cnst" never to
be chosen as the winner.
}
\examples{
conc = list(.03,.1,.3,1,3,10,30,100)
resp = list(0,.2,.1,.4,.7,.9,.6, 1.2)
row = list(conc = conc, resp = resp, bmed = 0, cutoff = 1, onesd = .5)
concRespCore(row, conthits = TRUE)
concRespCorec(row, aicc = TRUE)
}
