% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concRespPlot2.R
\name{concRespPlot2}
\alias{concRespPlot2}
\title{Concentration Response Plot - ggplot2}
\usage{
concRespPlot2(row, log_conc = FALSE)
}
\arguments{
\item{row}{Output from `concRespCore` or `tcplhit2_core`, containing information
about the best fit curve (winning model).}

\item{log_conc}{Logical argument. If `TRUE`, convert the concentrations (x-axis)
into log-10 scale. Defaults to `FALSE`.}
}
\value{
A `ggplot` object of the observed concentration-response data
overlaid with the best fit curve (winning model).
}
\description{
This function takes output from `concRespCore` or `tcplhit2_core` to
generate a basic plot of the observed concentration-response data and the
best fit curve (winning model). A `ggplot` object, which users may customize
with additional `ggplot` layers, is returned.
}
