% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AUC.R
\name{get_AUC}
\alias{get_AUC}
\title{Calculate AUC}
\usage{
get_AUC(fit_method, lower, upper, ps)
}
\arguments{
\item{fit_method}{name of the model}

\item{lower}{lower bound for x}

\item{upper}{upper bound for x}

\item{ps}{model parameter vector}
}
\value{
AUC value
}
\description{
Function that calculates AUC for dose-response curves
}
\details{
This function takes in a type of model and a set of model parameters,
and return the Area under the curve (AUC) between the lower and upper
concentration bounds. AUC can be used to compute an efficacy/ potency
metric for "active" does-response curves. The model parameters should be
entered as a list or a vector. For models operate in log scale (hill and
gain-loss), the lower and upper concentration bounds, "ga" (gain AC50)
and "la" (loss AC50) will be converted to log-scale if applicable.
This version hasn't taken consideration into how to appropriately work with
decreasing curves. One of the future works to do.
}
\examples{
conc <- c(.03,.1,.3,1,3,10,30,100)
fit_method <- "gnls"
modpars <- list(tp = 1.023, ga = 2.453, p = 1.592,
                la = 4288.993, q = 5.770, er = -3.295)
get_AUC("exp2", min(conc), max(conc), ps = modpars)

}
