% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplObj.R
\name{poly1}
\alias{poly1}
\title{Polynomial 1 Model}
\usage{
poly1(ps, x)
}
\arguments{
\item{ps}{Vector of parameters: a,er}

\item{x}{Vector of concentrations (regular units)}
}
\value{
Vector of model responses
}
\description{
Polynomial 1 Model
}
