% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplObj.R
\name{poly2}
\alias{poly2}
\title{Polynomial 2 Model}
\usage{
poly2(ps, x)
}
\arguments{
\item{ps}{Vector of parameters: a,b,er}

\item{x}{Vector of concentrations (regular units)}
}
\value{
Vector of model responses
}
\description{
Polynomial 2 Model
}
