% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplObj.R
\name{poly2bmds}
\alias{poly2bmds}
\title{Polynomial 2 Model (BMDS)}
\usage{
poly2bmds(ps, x)
}
\arguments{
\item{ps}{Vector of parameters: b1,b2,er}

\item{x}{Vector of concentrations (regular units)}
}
\value{
Vector of model responses
}
\description{
\eqn{f(x) = b1*x + b2*x^{2}}
}
\examples{
poly2bmds(c(1,2),1)

}
