% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplObj.R
\name{pow}
\alias{pow}
\title{Power Model - ps[1]*x^ps[2]}
\usage{
pow(ps, x)
}
\arguments{
\item{ps}{Vector of parameters: a,p,er}

\item{x}{Vector of concentrations (regular units)}
}
\value{
Vector of model responses
}
\description{
Power Model - ps[1]*x^ps[2]
}
\examples{
pow(c(1,2),1)

}
