% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplFit2.test.R
\name{tcplFit2.test}
\alias{tcplFit2.test}
\title{Run a test of the tcploFit2 code}
\usage{
tcplFit2.test()
}
\description{
This is jsut a test routine to show how tcplFit2 runs
}
\details{
concRespCore(row, conthits = TRUE)
concRespCorec(row, aicc = TRUE)
}
