% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplObj.R
\name{exp2}
\alias{exp2}
\title{Exponential 2 Model - ps[1]*(exp(x/ps[2]) - 1)}
\usage{
exp2(ps, x)
}
\arguments{
\item{ps}{Vector of parameters: a,b,er}

\item{x}{Vector of concentrations (regular units)}
}
\value{
Vector of model responses
}
\description{
Exponential 2 Model - ps[1]*(exp(x/ps[2]) - 1)
}
\examples{
exp2(c(1,2),1)

}
