% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplObj.R
\name{exp3}
\alias{exp3}
\title{Exponential 3 Model - ps[1]*(exp((x/ps[2])^ps[3]) - 1)}
\usage{
exp3(ps, x)
}
\arguments{
\item{ps}{Vector of parameters: a,b,p,er}

\item{x}{Vector of concentrations (regular units)}
}
\value{
Vector of model responses
}
\description{
Exponential 3 Model - ps[1]*(exp((x/ps[2])^ps[3]) - 1)
}
\examples{
exp3(c(1,2,2),1)

}
