% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplObj.R
\name{exp4}
\alias{exp4}
\title{Exponential 4 Model}
\usage{
exp4(ps, x)
}
\arguments{
\item{ps}{Vector of parameters: tp,ga,er}

\item{x}{Vector of concentrations (regular units)}
}
\value{
Vector of model responses
}
\description{
\eqn{f(x) = tp*(1-2^{(-x/ga)})}
}
\examples{
exp4(c(1,2),1)

}
