% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplObj.R
\name{exp5}
\alias{exp5}
\title{Exponential 5 Model - ps[1]*(1-2^(-(x/ps[2])^ps[3]))}
\usage{
exp5(ps, x)
}
\arguments{
\item{ps}{Vector of parameters: tp,ga,p,er}

\item{x}{Vector of concentrations (regular units)}
}
\value{
Vector of model responses
}
\description{
Exponential 5 Model - ps[1]*(1-2^(-(x/ps[2])^ps[3]))
}
\examples{
exp5(c(1,2,3),1)

}
