% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AUC.R
\name{get_AUC}
\alias{get_AUC}
\title{Calculate AUC}
\usage{
get_AUC(fit_method, lower, upper, ps)
}
\arguments{
\item{fit_method}{name of the winning model}

\item{lower}{lower bound for x}

\item{upper}{upper bound for x}

\item{ps}{model parameter vector}

\item{...}{placeholder, should be fine without it}
}
\value{
AUC value
}
\description{
Function that calculates AUC for dose-response curves
}
\details{
Assume for now we want to calculate AUC for the winning model.
The first half of tcplhit2_core selects a winning model and extract the model
name, model parameters from model fitting result (result from
tcplfit2_core) and populate them into the environment. We will have
fit_method - name of the winning model (string) and modpars - a list of
model parameters.
}
\examples{
conc <- c(.03,.1,.3,1,3,10,30,100)
modpars <- c(1, 2, er = -3.295307) #a, b, er
get_AUC("exp2", min(conc), max(conc), ps = modpars)

}
