% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplObj.R
\name{gnls}
\alias{gnls}
\title{Gain-Loss Model - ps[1] * (1/(1 + (ps[2]/x)^ps[3])) * (1/(1 + (x/ps[4])^ps[5]))}
\usage{
gnls(ps, x)
}
\arguments{
\item{ps}{Vector of parameters: tp,ga,p,la,q,er}

\item{x}{Vector of concentrations (regular units)}
}
\value{
Vector of model responses
}
\description{
Gain-Loss Model - ps[1] * (1/(1 + (ps[2]/x)^ps[3])) * (1/(1 + (x/ps[4])^ps[5]))
}
\examples{
gnls(c(1,2,1,2,2),1)

}
