% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplObj.R
\name{gnls}
\alias{gnls}
\title{Gain-Loss Model - \eqn{f(x) = \frac{tp}{[(1 + (ga/x)^p )(1 + (x/la)^q )]}}}
\usage{
gnls(ps, x)
}
\arguments{
\item{ps}{Vector of parameters: tp,ga,p,la,q,er}

\item{x}{Vector of concentrations (regular units)}
}
\value{
Vector of model responses
}
\description{
Gain-Loss Model - \eqn{f(x) = \frac{tp}{[(1 + (ga/x)^p )(1 + (x/la)^q )]}}
}
\examples{
gnls(c(1,2,1,2,2),1)

}
