% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplObj.R
\name{hillfn}
\alias{hillfn}
\title{Hill Model}
\usage{
hillfn(ps, x)
}
\arguments{
\item{ps}{Vector of parameters: tp,ga,p,er}

\item{x}{Vector of concentrations (regular units)}
}
\value{
Vector of model responses
}
\description{
\eqn{f(x) = \frac{tp}{[(1 + (ga/x)^p )]}}
}
\examples{
hillfn(c(1,2,3),1)

}
