% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplObj.R
\name{hillfn}
\alias{hillfn}
\title{Hill Model - ps[1]/(1 +  (ps[2]/x)^ps[3])}
\usage{
hillfn(ps, x)
}
\arguments{
\item{ps}{Vector of parameters: tp,ga,p,er}

\item{x}{Vector of concentrations (regular units)}
}
\value{
Vector of model responses
}
\description{
Hill Model - ps[1]/(1 +  (ps[2]/x)^ps[3])
}
\examples{
hillfn(c(1,2,3),1)

}
