% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplObj.R
\name{loggnls}
\alias{loggnls}
\title{Log Gain-Loss Model}
\usage{
loggnls(ps, x)
}
\arguments{
\item{ps}{Vector of parameters: tp,ga,p,la,q,er (ga and la are in log10-scale)}

\item{x}{Vector of concentrations (log10 units)}
}
\value{
Vector of model responses
}
\description{
\eqn{f(x) = \frac{tp}{[(1 + 10^{(p*(ga-x))} )(1 + 10^{(q*(x-la))} )]}}
}
\examples{
loggnls(c(1,2,1,2,2),1)

}
