% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplObj.R
\name{poly2}
\alias{poly2}
\title{Polynomial 2 Model - \eqn{f(x) = a*(\frac{x}{b} + \frac{x^2}{b^2})}}
\usage{
poly2(ps, x)
}
\arguments{
\item{ps}{Vector of parameters: a,b,er}

\item{x}{Vector of concentrations (regular units)}
}
\value{
Vector of model responses
}
\description{
Polynomial 2 Model - \eqn{f(x) = a*(\frac{x}{b} + \frac{x^2}{b^2})}
}
\examples{
poly2(c(1,2),1)

}
