% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hit_AUC.R
\name{post_hit_AUC}
\alias{post_hit_AUC}
\title{Calculate AUC After Hit-calling}
\usage{
post_hit_AUC(hit_results)
}
\arguments{
\item{hit_results}{output from tcplhit2_core}
}
\value{
AUC value of the winning model
}
\description{
Function that calculates AUC for dose-response curves after hit-calling
}
\details{
This function calculates AUC value for the winning model selected after
hit-calling.
}
\examples{

conc <- c(.03, .1, .3, 1, 3, 10, 30, 100)
resp <- c(0, .2, .1, .4, .7, .9, .6, 1.2)
params <- tcplfit2_core(conc, resp, .8)
output <- tcplhit2_core(params, conc, resp, 0.8, 0.5)
post_hit_AUC(output)


}
